+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Elgg Cropper - CHANGE LOG
 ========================
 
+### 2.1.2 (2016-12-07)
+* Fixed: removed unused graphics folder
+
 ### 2.1.1 (2016-11-15)
 * Plenty of Cropper tool options are now available on plugin settings
 * Added option to select size of preview box in settings